requirements.txt
requests
PyYAML
pystray
pillow
ttkbootstrap
